Got Tomcat working
