+++ .idea/shelf/Daily_updates.xml
@@ -1,0 +1,4 @@
+<changelist name="Daily_updates" date="1492207189850" recycled="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Daily_updates/shelved.patch" />
+  <option name="DESCRIPTION" value="Daily updates" />
+</changelist>
