+++ docker/Dockerfile
@@ -1,0 +1,9 @@
+FROM tomcat
+
+MAINTAINER "Rick MacKinnon <rick.j.mackinnon@gmail.com>"
+
+ADD settings.xml /usr/local/tomcat/conf/
+ADD tomcat-users.xml /usr/local/tomcat/conf/
+ADD context.xml /usr/local/tomcat/webapps/manager/META-INF
+ADD context.xml /usr/local/tomcat/webapps/host-manager/META-INF
+
