+++ docker/settings.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<settings>
+ <servers> 
+  <server>
+   <id>TomcatServer</id>
+   <username>rick</username>
+   <password>password</password>
+  </server> 
+ </servers>
+</settings>
+
