+++ docker/tomcat-users.xml
@@ -1,0 +1,12 @@
+
+<?xml version='1.0' encoding='utf-8'?>
+<tomcat-users>
+ <role rolename="manager"/>
+ <role rolename="manager-gui"/>
+ <role rolename="manager-script"/>
+ <role rolename="manager-jmx" />
+ <role rolename="admin-gui"/>
+ <user username="rick" password="password" roles="manager,manager-gui,admin-gui,standard" />
+ <user username="script" password="script" roles="manager-script,manager-jmx,standard" />
+</tomcat-users>
+
