--- src.main.java/com/magichamster/grocerysamurai/controller/CommonController.java
@@ -1,7 +1,0 @@
-package com.magichamster.grocerysamurai.controller;
-
-/**
- * Created by rick on 4/29/17.
- */
-public class CommonController {
-}
