--- src.main.java/main/resources/META-INF/persistence.xml
@@ -1,21 +1,0 @@
-<persistence xmlns="http://java.sun.com/xml/ns/persistence"
-             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd
-http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence"
-             version="2.0">
-    <persistence-unit name="grocery">
-        <provider>org.hibernate.jpa.HibernatePersistenceProvider</provider>
-        <!-- Annotated entity classes -->
-        <class>com.magichamster.grocerysamurai.model.Item</class>
-        <properties>
-
-            <property name="hibernate.connection.url" value="jdbc:postgresql://localhost:5432/grocery"/>
-            <property name="hibernate.connection.driver_class" value="org.postgresql.Driver"/>
-            <property name="hibernate.connection.username" value="postgres"/>
-            <property name="hibernate.connection.password" value="MyDbpassw0rd!"/>
-
-            <property name="hibernate.dialect" value="org.hibernate.dialect.PostgreSQL82Dialect"/>
-            <property name="hibernate.show_sql" value="true"/>
-        </properties>
-    </persistence-unit>
-</persistence>
