--- src.test.java/com/magichamster/grocerysamurai/repository/test/ItemRepositoryTest.java
@@ -1,23 +1,0 @@
-package com.magichamster.grocerysamurai.repository;
-
-import com.magichamster.grocerysamurai.model.Item;
-
-import javax.persistence.Persistence;
-
-/**
- * Test specifically for Item
- * Created by Rick on 4/15/17.
- */
-public class ItemRepositoryTest extends IBaseRepositoryTest<Item> {
-    @Override
-    public IBaseRepository<Item> getRepository() {
-        return new BaseRepository<>(Item.class, Persistence.createEntityManagerFactory("grocery"));
-    }
-
-    @Override
-    public Item createEntity(int id) {
-        Item entity = new Item();
-        entity.setName("Item" + id);
-        return entity;
-    }
-}
